Huge Commit
+ Fixed Chatlog Saving (And Chatlog Duplicates)
+ Fixed Pets Outgoing Message
+ Fixed Pets Save Data
+ Fixed Save User Stats
+ Improved Database Manager
+ Improved a lot of Code
+ Cleaned a Lot of Code
+ Other Stuff

+++ Yupi/Game/Rooms/Data/DynamicRoomModel.cs
@@ -242,7 +242,7 @@ namespace Yupi.Game.Rooms.Data
         {
             ServerMessage serverMessage = new ServerMessage(LibraryParser.OutgoingRequest("FloorMapMessageComposer"));
             serverMessage.AppendBool(true);
-            serverMessage.AppendInteger((int) _mRoom.RoomData.WallHeight);
+            serverMessage.AppendInteger(_mRoom.RoomData.WallHeight);
             StringBuilder stringBuilder = new StringBuilder();
             for (int i = 0; i < MapSizeY; i++)
             {
